get_default_global_state in derive_concentrations
