+++ vivarium/processes/derive_concentrations.py
@@ -1,6 +1,17 @@
 from vivarium.core.process import Deriver
-from vivarium.processes.derive_counts import get_default_global_state
+from vivarium.library.units import units
+from vivarium.processes.tree_mass import AVOGADRO
 
+def get_default_global_state():
+    mass = 1000 * units.fg
+    density = 1100 * units.g / units.L
+    volume = mass / density
+    mmol_to_counts = (AVOGADRO * volume)
+    return {
+        'global': {
+            'volume': volume.to('fL'),
+            'mmol_to_counts': mmol_to_counts.to('L/mmol')
+        }}
 
 class DeriveConcentrations(Deriver):
     """
